backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.9.2
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
honcho==1.1.0
idna==2.8
importlib_metadata==7.2.1
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.22.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==1.25.11
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
